:fire: Removed PseudoState, FinalState and Vertex
Removed PseudoState and FinalState as they can be simulated by just moving a an attribute of type PseudoStateKind into State. For that PseudoKindState got two new constants "STATE" representing a "normal" state and "FINAL" representing a FinalState. Thus the need for the generalization vertex ceased to exist and it was fused into State. State in turn can now extend Namespace instead. The changes were ported to LookupUtil