:recycle: IRestructurer changed contract + CachingRestructurer
A restructurer now only offers two methods, to restructure an entire model and to restructure a single element. Instead of having a specialized version for noncaching IRestructurers there is now a specialization for caching ones. The map used as a cache is not static anymore and thus could be extracted to the general class wich cleans up a lot of documentation duplication