:truck: Made package restructure subpackage of inputprocessing
In one of the following commits there are going to be new classes that allow for easier switching of the input processing scheme. The current codegenerator is dependent on XMI and especially only UML as a modelling language. Instead of forcing the use of these two standards developers may implement the interfaces and use their own DSL or whatever modelling tools they desire