:new: Added an abstraction for oo-classes and other concepts
The goal of this more abstract concept is to move away from the uml implementation and use this instead. The main benefit is the increased ease with which one can change the input processing chain of the codegenerator. Developers only need to mold their classes into those OO-based ones instead of using the uml-faithful implementation as a step in between