:wrench: PrimitiveTypes: fixed issue with type being null
Primitive types in xmi are given by referencing a special xmi file. Thats the reason why up until now they caused an exception when using an advanced template. This was fixed by relying on on the part after # to identify the primitive type. 
Furthermore some transformers got a cache to fix the issue of a infinite loop